Refactor NotFoundException
- extend MyNakadiRuntimeException1
- move to the correct package
