+++ src/main/java/org/zalando/nakadi/exceptions/runtime/NotFoundException.java
@@ -1,6 +1,6 @@
-package org.zalando.nakadi.exceptions;
+package org.zalando.nakadi.exceptions.runtime;
 
-public class NotFoundException extends RuntimeException {
+public class NotFoundException extends MyNakadiRuntimeException1 {
 
     public NotFoundException(final String message) {
         super(message);
